Convert TransitionTable to Transitions page
